+++ packages/embed-wallet/src/index.ts
@@ -4,4 +4,6 @@ export { preloadIframe as prefetchWalletIframe } from '@cere/torus-embed';
 export * from './EmbedWallet';
 export * from './types';
 export { WALLET_CLIENT_VERSION } from './constants';
-export type { ProviderInterface } from './Provider';
+
+export type { SignerOptions } from './Signer';
+export type { ProviderInterface, SignerInterface } from './Provider';