+++ packages/wallet-engine/src/accounts.ts
@@ -3,6 +3,7 @@ import { getED25519Key } from '@toruslabs/openlogin-ed25519';
 import { decodeAddress, encodeAddress, isEthereumAddress } from '@polkadot/util-crypto';
 import { hexToU8a, isHex } from '@polkadot/util';
 import { Keyring } from '@polkadot/keyring';
+import { Keypair as SolKeypair } from '@solana/web3.js';
 
 import { KeyPair, KeyType, Account } from './types';
 import { CERE_SS58_PREFIX } from './constants';
@@ -29,6 +30,18 @@ const pairFactoryMap: Record<KeyType, (privateKey: string) => KeyPair> = {
       address: encodeAddress(publicKey, CERE_SS58_PREFIX),
     };
   },
+
+  solana: (privateKey) => {
+    const { sk: ed25519Key } = getED25519Key(privateKey);
+    const { publicKey, secretKey } = SolKeypair.fromSecretKey(ed25519Key);
+
+    return {
+      type: 'solana',
+      publicKey: publicKey.toBuffer(),
+      secretKey: Buffer.from(secretKey),
+      address: publicKey.toBase58(),
+    };
+  },
 };
 
 export type KeyPairOptions = {
@@ -44,17 +57,32 @@ export const getKeyPair = ({ privateKey, type }: KeyPairOptions): KeyPair => {
   return pairFactoryMap[type](privateKey);
 };
 
+export const getKeyPairs = (privateKey: string, types = Object.keys(pairFactoryMap) as KeyType[]) => {
+  return types.map((type) => getKeyPair({ privateKey, type }));
+};
+
 export const exportAccountToJson = ({ privateKey, type, passphrase }: KeyPairOptions & { passphrase?: string }) => {
+  if (type === 'solana') {
+    throw new Error('Not implemented');
+  }
+
   const { publicKey, secretKey } = getKeyPair({ type, privateKey });
   const keyring = new Keyring({ type });
 
   return keyring.addFromPair({ publicKey, secretKey }).toJson(passphrase);
 };
 
-export const getAccount = ({ privateKey, type, name }: AccountOptions): Account => ({
+export const getAccount = ({ privateKey, type, name }: AccountOptions): Account => {
+  const pair = getKeyPair({ privateKey, type });
+
+  return createAccountFromPair(pair, name);
+};
+
+export const createAccountFromPair = ({ type, address, publicKey }: KeyPair, name: string): Account => ({
   type,
   name,
-  address: getKeyPair({ privateKey, type }).address,
+  address,
+  publicKey: publicKey.toString('hex'),
 });
 
 const isValidPolkadotAddress = (address: string) => {
@@ -67,5 +95,10 @@ const isValidPolkadotAddress = (address: string) => {
   }
 };
 
-export const isValidAddress = (address: string, type: KeyType) =>
-  type === 'ethereum' ? isEthereumAddress(address) : isValidPolkadotAddress(address);
+export const isValidAddress = (address: string, type: KeyType) => {
+  if (type === 'solana') {
+    throw new Error('Not implemented');
+  }
+
+  return type === 'ethereum' ? isEthereumAddress(address) : isValidPolkadotAddress(address);
+};