+++ packages/wallet-engine/src/engine/accounts.ts
@@ -1,59 +1,71 @@
 import { createScaffoldMiddleware, createAsyncMiddleware } from 'json-rpc-engine';
 
 import { Engine } from './engine';
-import { getKeyPair } from '../accounts';
+import { getKeyPairs } from '../accounts';
 import { Account, KeyPair, KeyType } from '../types';
 
 export type AccountsEngineOptions = {
-  getAccounts: (pairs: KeyPair[]) => Account[];
+  getAccounts: () => Account[];
   getPrivateKey: () => string | undefined;
-  onUpdateAccounts: (accounts: Account[]) => void;
+  onUpdateAccounts: (accounts: KeyPair[]) => void;
 };
 
 export const createAccountsEngine = ({ getPrivateKey, getAccounts, onUpdateAccounts }: AccountsEngineOptions) => {
   const engine = new Engine();
 
-  const createAccounts = (types: KeyType[]) => {
-    const privateKey = getPrivateKey();
-
-    return !privateKey ? [] : getAccounts(types.map((type) => getKeyPair({ type, privateKey })));
-  };
+  const getAddresses = (type: KeyType) =>
+    getAccounts()
+      .filter((account) => account.type === type)
+      .map((account) => account.address);
 
   engine.push(
     createScaffoldMiddleware({
       wallet_accounts: createAsyncMiddleware(async (req, res) => {
-        res.result = createAccounts(['ethereum', 'ed25519']);
+        res.result = getAccounts();
       }),
 
       ed25519_accounts: createAsyncMiddleware(async (req, res) => {
-        res.result = createAccounts(['ed25519']);
+        res.result = getAddresses('ed25519');
       }),
 
       eth_accounts: createAsyncMiddleware(async (req, res) => {
-        res.result = createAccounts(['ethereum']).map((account) => account.address);
+        res.result = getAddresses('ethereum');
+      }),
+
+      solana_accounts: createAsyncMiddleware(async (req, res) => {
+        res.result = getAddresses('solana');
       }),
 
       eth_requestAccounts: createAsyncMiddleware(async (req, res) => {
-        res.result = createAccounts(['ethereum']).map((account) => account.address);
+        res.result = getAddresses('ethereum');
       }),
 
       wallet_updateAccounts: createAsyncMiddleware(async (req, res) => {
-        const accounts = createAccounts(['ethereum', 'ed25519']);
-        const [eth, ed255519] = accounts;
+        const privateKey = getPrivateKey();
+        const keyPairs = privateKey ? getKeyPairs(privateKey, ['ethereum', 'ed25519', 'solana']) : [];
 
-        onUpdateAccounts(accounts);
+        onUpdateAccounts(keyPairs);
+
+        const accounts = getAccounts();
+        const ethAccounts = accounts.filter((account) => account.type === 'ethereum');
+        const ed25519Accounts = accounts.filter((account) => account.type === 'ed25519');
+        const solanaAccounts = accounts.filter((account) => account.type === 'solana');
 
         /**
          * Custom wallet messages
          */
         engine.emit('message', { type: 'wallet_accountsChanged', data: accounts });
-        engine.emit('message', { type: 'eth_accountChanged', data: eth });
-        engine.emit('message', { type: 'ed25519_accountChanged', data: ed255519 });
+        engine.emit('message', { type: 'eth_accountsChanged', data: ethAccounts });
+        engine.emit('message', { type: 'ed25519_accountsChanged', data: ed25519Accounts });
+        engine.emit('message', { type: 'solana_accountsChanged', data: solanaAccounts });
 
         /**
          * Standard eip-1193 event
          */
-        engine.emit('accountsChanged', eth ? [eth.address] : []);
+        engine.emit(
+          'accountsChanged',
+          ethAccounts.map((account) => account.address),
+        );
 
         res.result = accounts;
       }),
@@ -61,7 +73,7 @@ export const createAccountsEngine = ({ getPrivateKey, getAccounts, onUpdateAccou
       wallet_getProviderState: createAsyncMiddleware(async (req, res, next) => {
         res.result = {
           ...(res.result as {}),
-          accounts: createAccounts(['ethereum']).map((account) => account.address),
+          accounts: getAddresses('ethereum'),
         };
       }),
     }),