+++ src/components/Permissions/knownPermissions.ts
@@ -16,6 +16,11 @@ export const knownPermissions: Record<AllowedPermission, PermissionMeta> = {
     title: 'Sign message (Cere Network)',
     description: 'Allow the app to sign data on your behalf in background.',
   },
+
+  solana_signMessage: {
+    title: 'Sign message (Solana)',
+    description: 'Allow the app to sign data on your behalf in background.',
+  },
 };
 
 export const defaultDescription = 'No description available';