+++ src/routes/Settings/Settings.tsx
@@ -10,6 +10,7 @@ import {
   Typography,
   styled,
   useIsMobile,
+  TextField,
 } from '@cere-wallet/ui';
 import { useEffect, useState } from 'react';
 
@@ -24,8 +25,23 @@ const SectionHeader = styled(CardHeader)({
 
 const SectionButton = styled(Button)({
   minWidth: 250,
-  height: 44,
+  whiteSpace: 'nowrap',
+  height: 42,
 }) as typeof Button;
+
+const downloadFile = (url: string, filename: string) => {
+  const link = document.createElement('a');
+
+  link.href = url;
+  link.download = filename;
+
+  document.body.appendChild(link);
+  link.click();
+  document.body.removeChild(link);
+
+  // Clean up by revoking the Blob URL
+  URL.revokeObjectURL(url);
+};
 
 export const Settings = () => {
   const isMobile = useIsMobile();
@@ -33,8 +49,13 @@ export const Settings = () => {
   const authenticationStore = useAuthenticationStore();
   const { accountUrl } = useOpenLoginStore();
   const [accountLink, setAccountLink] = useState<string>();
-  const walletDownloadUrl = accountStore.exportAccount('ed25519');
   const cereAddress = accountStore.getAccount('ed25519')?.address;
+  const [exportPassword, setExportPassword] = useState('');
+
+  const handleExportAccount = () => {
+    downloadFile(accountStore.exportAccount('ed25519', exportPassword), `${cereAddress}.json`);
+    setExportPassword('');
+  };
 
   useEffect(() => {
     authenticationStore
@@ -74,7 +95,7 @@ export const Settings = () => {
 
         <Card>
           <SectionHeader
-            title="Account export"
+            title="Export Your Account as an Encrypted JSON File"
             avatar={
               <IconButton variant="filled" size="medium">
                 <DownloadIcon />
@@ -82,21 +103,37 @@ export const Settings = () => {
             }
           />
           <CardContent>
-            <Stack direction={isMobile ? 'column' : 'row'} spacing={3}>
+            <Stack spacing={1}>
               <Typography flex={1} variant="body2" color="text.secondary">
-                JSON backup file lets you restore your account or use it with Cere Tools, even if a direct connection to
-                Cere Wallet isn't available. Please keep it confidential and don't share it with third parties.
+                This downloadable file lets you restore your account or use it with Cere Tools, even if you can't
+                connect directly to Cere Wallet. To ensure maximum security, the file will be encrypted with a password
+                you create.
               </Typography>
 
-              <SectionButton
-                disabled={!cereAddress}
-                download={`${cereAddress}.json`}
-                fullWidth={isMobile}
-                href={walletDownloadUrl}
-                variant="outlined"
-              >
-                Download
-              </SectionButton>
+              <Typography flex={1} variant="body2" color="text.secondary">
+                Please keep your password confidential and do not share it with anyone. Sharing your password could
+                compromise your account security.
+              </Typography>
+
+              <Stack direction={isMobile ? 'column' : 'row'} spacing={2} paddingTop={2}>
+                <TextField
+                  label="Encryption Password"
+                  value={exportPassword}
+                  fullWidth
+                  size="small"
+                  type="password"
+                  onChange={(event) => setExportPassword(event.target.value)}
+                />
+
+                <SectionButton
+                  disabled={!cereAddress || !exportPassword}
+                  fullWidth={isMobile}
+                  variant="contained"
+                  onClick={handleExportAccount}
+                >
+                  Export Account
+                </SectionButton>
+              </Stack>
             </Stack>
           </CardContent>
         </Card>