+++ CHANGELOG.md
@@ -2,7 +2,12 @@
 
 ### vNext
 
-- 
+-
+
+### v1.38.0
+
+- Added multi-chain signature API to the wallet client and SDK
+- Let users enter an encryption password during account export
 
 ### v1.37.0